chore: remove deprecated versions of React Router v6
