+++ .changeset/sour-seas-ring.md
@@ -1,0 +1,5 @@
+---
+'@backstage/core-app-api': patch
+---
+
+Remove deprecated versions of React Router v6.
