+++ packages/core-app-api/package.json
@@ -72,11 +72,8 @@
     "msw": "^1.0.0",
     "react": "^18.0.2",
     "react-dom": "^18.0.2",
-    "react-router-beta": "npm:react-router@6.0.0-beta.0",
-    "react-router-dom": "^6.3.0",
-    "react-router-dom-beta": "npm:react-router-dom@6.0.0-beta.0",
-    "react-router-dom-stable": "npm:react-router-dom@^6.3.0",
-    "react-router-stable": "npm:react-router@^6.3.0"
+    "react-router": "^6.3.0",
+    "react-router-dom": "^6.3.0"
   },
   "peerDependencies": {
     "@types/react": "^17.0.0 || ^18.0.0",
