+++ packages/core-app-api/src/app/AppManager.tsx
@@ -62,7 +62,6 @@ import {
 import { pluginCollector } from '../plugins/collectors';
 import {
   featureFlagCollector,
-  routingV1Collector,
   routingV2Collector,
 } from '../routing/collectors';
 import { RoutingProvider } from '../routing/RoutingProvider';
@@ -83,7 +82,6 @@ import { AppThemeProvider } from './AppThemeProvider';
 import { defaultConfigLoader } from './defaultConfigLoader';
 import { ApiRegistry } from '../apis/system/ApiRegistry';
 import { resolveRouteBindings } from './resolveRouteBindings';
-import { isReactRouterBeta } from './isReactRouterBeta';
 import { InternalAppContext } from './InternalAppContext';
 import { AppRouter, getBasePath } from './AppRouter';
 import { AppLanguageSelector } from '../apis/implementations/AppLanguageApi';
@@ -246,23 +244,11 @@ export class AppManager implements BackstageApp {
       );
 
       const { routing, featureFlags } = useMemo(() => {
-        const usesReactRouterBeta = isReactRouterBeta();
-        if (usesReactRouterBeta) {
-          // eslint-disable-next-line no-console
-          console.warn(`
-DEPRECATION WARNING: React Router Beta is deprecated and support for it will be removed in a future release.
-                     Please migrate to use React Router v6 stable.
-                     See https://backstage.io/docs/tutorials/react-router-stable-migration
-`);
-        }
-
         const result = traverseElementTree({
           root: children,
           discoverers: [childDiscoverer, routeElementDiscoverer],
           collectors: {
-            routing: usesReactRouterBeta
-              ? routingV1Collector
-              : routingV2Collector,
+            routing: routingV2Collector,
             collectedPlugins: pluginCollector,
             featureFlags: featureFlagCollector,
           },
