+++ packages/core-app-api/src/routing/collectors.test.tsx
@@ -32,10 +32,8 @@ import {
 } from '@backstage/core-plugin-api';
 import { MemoryRouter, Routes, Route } from 'react-router-dom';
 
-jest.mock('react-router', () => jest.requireActual('react-router-stable'));
-jest.mock('react-router-dom', () =>
-  jest.requireActual('react-router-dom-stable'),
-);
+jest.mock('react-router', () => jest.requireActual('react-router'));
+jest.mock('react-router-dom', () => jest.requireActual('react-router-dom'));
 
 const MockComponent = ({ children }: PropsWithChildren<{ path?: string }>) => (
   <>{children}</>
